Changed typdef for TypeTag to uintptr_t.
This makes it easier to use the type tag both as a tag and a pointer to metadata about the type.

+++ src/gperftools/typed_tcmalloc.h
@@ -26,8 +26,9 @@
 
 
 #include <stddef.h>                     /* for size_t */
+#include <stdint.h>                     /* for uintptr_t */
 
-typedef unsigned long TypeTag;
+typedef uintptr_t TypeTag;
 
 #ifdef __cplusplus
 extern "C" {
